% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{ctcf_rep2}
\alias{ctcf_rep2}
\title{CTCF rep2}
\format{A GRanges object with an mcols DataFrame with columns:
\describe{
  \item{mcols.signal}{signal}
  \item{mcols.pValue}{the pValue of the peak}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgFileUi?db=hg19&g=wgEncodeUwTfbs}
}
\description{
CTCF ChIP-Seq peak intensity in MCF-7 cells, replicate 2.
}

