\name{fmdatamcf7parp1-package}
\alias{fmdatamcf7parp1-package}
\alias{fmdatamcf7parp1}
\docType{package}
\title{
Fondufe-Mittendorfe lab data for PARP1 binding in MCF-7
}
\description{
Contains the data used for analysis of PARP1 nucleosome binding in MCF-7 cells by the Fondufe-Mittendorf lab at University of Kentucky
}
\details{
\tabular{ll}{
Package: \tab fmdatamcf7parp1\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-12-04\cr
License: \tab CC0\cr
}

}
\author{
Robert M Flight

Maintainer: Robert M Flight <rflight79@gmail.com>
}