% Generated by roxygen2 (4.0.2): do not edit by hand
\name{histone_marks}
\alias{histone_marks}
\title{Histone Marks}
\format{GRangesList of length 6, with H3k09me3, H3k27ac, H3k27me3, H3k36me3,
H3k4me3_r1 and H3k4me3_r2. Each is a GRanges object with mcols of:
\describe{
  \item{mcols.score}{the score}
  \item{mcols.signal}{the peak intensity}
  \item{mcols.pvalue}{-1 * log10 of the pvalue}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTracks?hgsid=339462435&hgt_=1371663888&db=hg19&tsCurTab=advancedTab&hgt_tsDelRow=&hgt_tsAddRow=&hgt_tsPage=&tsSimple=&tsName=histone&tsDescr=&tsGroup=Any&tsType=Any&hgt_mdbVar1=cell&hgt_mdbVal1=MCF-7&hgt_mdbVar2=antibody&hgt_mdbVal2=Any&hgt_tSearch=search}
}
\description{
ChIP-Seq peaks of various histone marks
}

