% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{methyl_rep1}
\alias{methyl_rep1}
\title{Methylation rep1}
\format{A GRanges object with an mcols DataFrame with columns:
\describe{
  \item{mcols.readCount}{the number of reads}
  \item{mcols.percentMeth}{the percentage of reads that were methylated}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables?hgsid=396587911_8YHquTEUSQjmIfAHtJJT7vWY7N8U&clade=mammal&org=Human&db=hg19&hgta_group=allTracks&hgta_track=wgEncodeHaibMethylRrbs&hgta_table=wgEncodeHaibMethylRrbsMcf7DukeSitesRep1}
}
\description{
Methylation reads from MCF-7, replicate 1
}

