% Generated by roxygen2 (4.0.2): do not edit by hand
\name{methyl_rep2}
\alias{methyl_rep2}
\title{Methylation rep2}
\format{A GRanges object with an mcols DataFrame with columns:
\describe{
  \item{mcols.readCount}{the number of reads}
  \item{mcols.percentMeth}{the percentage of reads that were methylated}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables?hgsid=396587911_8YHquTEUSQjmIfAHtJJT7vWY7N8U&clade=mammal&org=Human&db=hg19&hgta_group=allTracks&hgta_track=wgEncodeHaibMethylRrbs&hgta_table=wgEncodeHaibMethylRrbsMcf7DukeSitesRep2}
}
\description{
Methylation reads from MCF-7, replicate 2
}

