% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{nuc_signal}
\alias{nuc_signal}
\title{MNase Nucleosome Data}
\format{Granges with an mcol called \code{nuc}, holding the summed signal over the
tile.}
\source{
\url{ftp://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeSydhNsome/wgEncodeSydhNsomeGm12878Sig.bigWig}
}
\description{
MNase nucleosome signal from GM12878, generated by ENCODE, summed over 500 base
tiles of the genome.
}

