% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parp1_ln4_reads_all}
\alias{parp1_ln4_reads_all}
\title{All the PARP1 reads from LN4}
\format{A GRanges object.}
\source{
Processing reads from YFM
}
\description{
A dataset containing all the PARP1 reads from LN4 (MCF-7). This data consists of three biological replicates
sequenced on a single Illumina lane.
}

