% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{parp1_ln4_unique}
\alias{parp1_ln4_unique}
\title{The unique location PARP1 reads from LN4}
\format{A GRanges object with an mcols DataFrame with a column named n_count}
\source{
Processing reads from YFM
}
\description{
A dataset containing only the unique location PARP1 reads from LN4 (MCF-7), with the number of reads
at each location capped at 6. The cap of 6 comes from having unique reads, in three biological
replicates, and ignoring strand information.
}

