% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parp1_ln5_unique}
\alias{parp1_ln5_unique}
\title{The unique location PARP1 reads from LN5}
\format{A GRanges object with an mcols DataFrame with a column named n_count}
\source{
Processing reads from YFM
}
\description{
A dataset containing only the unique location PARP1 reads from LN5 (MDA-MB231), with the number of reads
at each location capped at 6. The cap of 6 comes from having unique reads, in three biological
replicates, and ignoring strand information.
}

