% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{tss_windows}
\alias{tss_windows}
\title{Transcript start sites}
\format{A named GRanges object, where the names are the Ensembl transcript IDs}
\source{
ensembl transcripts
}
\description{
2KB windows around Ensembl transcript start sites (TSS)
}

