% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expr_data}
\alias{expr_data}
\title{Expression Data}
\format{data.frame with:
\describe{
  \item{ID_REF}{the affymetrix probe set id}
  \item{VALUE}{the normalized signale}
  \item{ABS_CALL}{absence presence call}
  \item{P-VALUE}{p-value of being present}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM307014}
}
\description{
DNA Microarray Expression Data from GEO, Affymetrix hgu133plus2, GSM307014
measured on MCF-7 cells
}

