% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmdatamcf7parp1.R
\name{get_chr}
\alias{get_chr}
\title{get chromosome}
\usage{
get_chr(filename, split_chr)
}
\arguments{
\item{filename}{the filename to parse}

\item{split_chr}{the character separating the chr* from the rest}
}
\value{
string of the chromosome
}
\description{
from a filename, get the chromosome part of the filename
}

