% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tss_windows}
\alias{tss_windows}
\title{Transcript start sites}
\format{A named GRanges object, where the names are the Ensembl transcript IDs}
\source{
ensembl transcripts
}
\description{
2KB windows around Ensembl transcript start sites (TSS)
}

